+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [0.0.5] - 2024-06-03
+
+### Added
+
+* Expanded Moore neighborhood.
+* Abstract model has agents indexed by type.
+
 ## [0.0.4] - 2024-06-02
 
 ### Added